backups/setup/server/requirements.txt
python-crontab
awscli
click